+++ lib/src/base_api.dart
@@ -6,35 +6,43 @@ part of http_api;
 /// Override this class in order to
 /// add your own request methods
 abstract class BaseApi {
-  final Uri _url;
-  Uri get url => _url;
+  final Uri url;
   final Map<String, String> defaultHeaders;
   final ApiLink _link;
 
   BaseApi(
-    Uri url, {
+    this.url, {
     ApiLink link,
     Map<String, String> defaultHeaders,
-  })  : assert(url != null, "url $runtimeType argument cannot be null"),
-        _url = url,
-        defaultHeaders = defaultHeaders ?? <String, String>{},
+  })  : defaultHeaders = defaultHeaders ?? <String, String>{},
         _link = link?._firstLink ?? HttpLink() {
+    ArgumentError.checkNotNull(url, 'url');
+
     if (_link._firstWhere((apiLink) => (apiLink is HttpLink)) == null) {
-      throw ApiError("ApiLinks chain should contain HttpLink");
+      throw ApiError("ApiLinks chain should contain HttpLink.");
     }
 
-    if (_link._firstWhere((apiLink) => apiLink.closed) != null) {
-      throw ApiError("Cannot assign closed ApiLinks chain to $runtimeType");
+    if (_link._firstWhere((apiLink) => apiLink.attached) != null) {
+      throw ApiError(
+        "Cannot reattach already attached ApiLink to $runtimeType.",
+      );
     }
 
     /// Close link chain
-    _link._closeChain();
+    _link._closeChain(url);
   }
 
   /// Get first link of provided type from current link chain.
   /// If ApiLink chain does not contain link of provided type,
   /// [null] will be returned.
-  T getFirstLinkOfType<T>() => _link?._firstWhere((link) => link is T) as T;
+  T getFirstLinkOfType<T extends ApiLink>() =>
+      _link?._firstWhere((link) => link is T) as T;
+
+  ApiLink getFirstLinkWhere(bool Function(ApiLink link) test) =>
+      _link?._firstWhere(test);
+
+  void forEachLink(void Function(ApiLink) callback) =>
+      _link?._forEach(callback);
 
   /// Make API request by triggering [ApiLink]s [next] methods
   Future<ApiResponse> send(ApiRequest request) async {
@@ -43,9 +51,6 @@ abstract class BaseApi {
     request.headers.addAll(
       Map<String, String>.from(defaultHeaders)..addAll(request.headers),
     );
-
-    /// Sets request api url
-    request._apiUrl = url;
 
     return _link.next(request);
   }